Mapi/requirements.txt
certifi==2019.6.16
Django==2.2.3
djangorestframework==3.10.2
joblib==0.13.2
numpy==1.17.0
pandas==0.25.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0